+++ server/src/game/commonConfig.ts
