--- server/src/game/model/item.ts
@@ -1,3 +1,0 @@
-import { GameObject } from './gameObject';
-
-export class Item extends GameObject {}
