--- server/src/game/model/npc.ts
@@ -1,5 +1,0 @@
-import { Character } from "./character";
-
-export class Npc extends Character{
-    
-}
