--- server/src/game/model/obstacle.ts
@@ -1,3 +1,0 @@
-import { GameObject } from './gameObject';
-
-export class Obstacle extends GameObject {}
