--- server/src/game/model/player.ts
@@ -1,5 +1,0 @@
-import { Character } from "./character";
-
-export class Player extends Character{
-    
-}
