+++ server/src/game/model/wall.ts
