+++ server/src/game/util/overlapTester.ts
@@ -1,8 +1,6 @@
-import { RectBound } from "../../types/rectBound.type";
-
 // 干渉するかテストする静的関数群クラス
 export class OverlapTester {
-  static overlapRects(rect1: RectBound, rect2: RectBound) {
+  static overlapRects(rect1: any, rect2: any) {
     // 矩形１の左端と矩形２の右端を比べて、矩形１の左端の方が右にあるなら、重ならない。
     if (rect1.left > rect2.right) {
       return false;
@@ -23,7 +21,7 @@ export class OverlapTester {
     return true;
   }
 
-  static pointInRect(rect: RectBound, point: any) {
+  static pointInRect(rect: any, point: any) {
     return (
       rect.left <= point.x &&
       rect.right >= point.x &&
